requirement.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==24.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
idna==3.11
PyJWT==2.11.0
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
